Have `Node` take a genric `KVStore` parameter
Previously, we had `Node` take a concrete `FilesystemStore` struct. Here
we switch to a generic `KVStore` type parameter. To this end we switched
from `K: Deref` to concrete `Arc<K>` in all of the modules to avoid
confusion of the type paramters or requirements to track `Arc`-ed and
non-`Arc`ed version of the `K: KVStore` paramter. Moreover, as Uniffi
doesn't support exposing generics we now expose a concretized `LDKNode`
type alias in bindings, which will use `SqliteStore` in the future.

Note that going the generic route was necessary as `dyn KVStore`
wasn't an opion due to the incompatibility of `dyn` with associated
types. In this case the incompatibility is in regard to
`KVStore::Reader` and I opted to go this route over forcing any
implementation of the trait to allocating and returning the same
concrete value (e.g., `Vec<u8>`) which could potentially have
significant performance impacts over returning a buffered reader for
example.