Drop `Node::onchain_balance` from public API
We already have `Node::spendable_onchain_balance` and
`Node::total_onchain_balance` exposed, so the general `onchain_balance`
is kind of redudant. As it is also not exposed (and not easily
exposable) in bindings, we can make the bindings and Rust APIs more
homogeneous by dropping the call.

Plus, we probably should avoid exposing any BDK types in the public API
until BDK 1.0 landed and we made the switch, as many breaking changes
will be coming with it.