Improve `sync_wallet` docs and use current thread runtime for LDK
We improve the docs, in particular mentioning now that users should
always prefer background syncing over blocking sync if possible.

Moreover, as we already have a current thread runtime setup for the BDK
wallet, we now also use it to sync the LDK wallet. This is a good idea
as we want to avoid calling `block_on` on our main runtime where
possible, which is always in danger of (temporarily or permanently)
blocking the world if something goes wrong.