Remove unnecessary `PaymentStatus::SendingFailed`
Previously, I had a misunderstood in what event retrying payments is
safe. It turns out that it's indeed safe to retry a payment as soon as
we receive `LdkEvent::PaymentFailed`.

Therefore, we can here now only refuse to send duplicate payments if our previous status is
`Succeeded` or `Pending`, which renders the additional distinction
between `Failed` and `SendingFailed` superfluous. Accordingly, we remove
this `PaymentStatus` variant.