Clean up Kotlin/JVM bindings generation script
Previously, the targets built by the script was dependent on the
platform on which it was run, also delegating some of the logic to the
CI publishing scripts.

Here, we make the built targets explicit (aarch64 and x86_64 on
macOS/darwin and x86_64 on Linux) and copy the artifacts over to their
respective ressource locations.