Fix logger symlink setup
Currently we check the for the existence of the log file the symlink is
pointing to before removing the symlink. This could lead to weird issues
when a log file was removed, as we'd then never remove and update the
symlink.

Here, we remove the superfluous check and just check that the path in
question is indeed a symlink before removing it.

While we're at it we introduce some logging to `stderr` that might help
debug issues arising during logger setup.