Manually download `bitcoind`/`electrs` for CI tests
Previously, we used the auto-download feature of the
`electrsd`/`bitcoind` crates. While convenient, they unnecessarily
introduced a lot of dependecies (`zip`, `zstd`, `time`, etc.) to our
test environment which needed pinning for the MSRV of 1.63.

Here, we introduce a new `no_download` config flag allowing us to
disable this auto-download feature in CI, where we now opt to download
the corresponding binaries manually. We keep the default-auto-download
as a convenience feature for running tests locally though.