Bump max inflight HTLC value to 100% if we're an LSPS2 client
When we're an LSPS2 client, the LSP might open the channel to us with a
capacity only barely more than the initial payment size. We therefore
need to be able to receive close to 100% of the channel size in a single
HTLC.

Here we therefore bump the default max HTLC value inflight config to
100%. We should eventually be able to set this on a per-channel basis,
but for now we just bump the default for all channels.