Reconnect to all persisted peers after restart
While we previously took care of persisting peer details when a channel
is opened and if the user signals to `connect` to have the connection
persisted, our background task would still only try to reconnect to
channel peers. Here, we fix this omission and add a regression test
ensuring we'll see the proper behavior going forward.