feat: sanitize and set node alias
What this commit does:
Implements a method `set_node_alias` on NodeBuilder to allow
callers customize/set the value of the node alias. This method
sanitizes the user-provided alias by ensuring the following:
 + Node alias is UTF-8-encoded String
 + Node alias is non-empty
 + Node alias cannot exceed 32 bytes
 + Node alias is only valid up to the first null byte. Every
   character after the null byte is discraded

Additionally, a test case is provided to cover sanitizing
empty node alias, as well as an alias with emojis (copied and
modified from rust-lightning) and a sandwiched null byte.