Implement `PayjoinPayment` ..
Payjoin [`BIP77`] implementation. Compatible with previous Payjoin version [`BIP78`].

Should be retrieved by calling [`Node::payjoin_payment`].

Payjoin transactions can be used to improve privacy by breaking the common-input-ownership
heuristic when Payjoin receivers contribute input(s) to the transaction. They can also be used
to save on fees, as the Payjoin receiver can direct the incoming funds to open a lightning
channel, forwards the funds to another address, or simply consolidate UTXOs.

In a Payjoin transaction, both the sender and receiver contribute inputs to the transaction in
a coordinated manner. The Payjoin mechanism is also called pay-to-endpoint(P2EP).  The Payjoin
receiver endpoint address is communicated through a [`BIP21`] URI, along with the payment
address and an optional amount parameter.  In the Payjoin process, parties edit, sign and pass
iterations of the transaction between each other, before a final version is broadcasted by the
Payjoin sender.  [`BIP77`] codifies a protocol with 2 iterations (or one round of interaction
beyond address sharing).

[`BIP77`] Defines the Payjoin process to happen asynchronously, with the Payjoin receiver
enrolling with a Payjoin Directory to receive Payjoin requests. The Payjoin sender can then
make requests through a proxy server, Payjoin Relay, to the Payjoin receiver even if the
receiver is offline. This mechanism requires the Payjoin sender to regularly check for response
from the Payjoin receiver as implemented in [`Node::payjoin_payment::send`].

A Payjoin Relay is a proxy server that forwards Payjoin requests from the Payjoin sender to the
Payjoin receiver subdirectory. A Payjoin Relay can be run by anyone. Public Payjoin Relay
servers are:
- https://pj.bobspacebkk.com

A Payjoin directory is a service that allows Payjoin receivers to receive Payjoin requests
offline. A Payjoin directory can be run by anyone. Public Payjoin Directory servers are:
- https://payjo.in