Refactor `max_total_routing_fee_msat` to allow setting `None`
Previously, the `SendingParamters` field was simply an `Option<u64>`,
which however means we could just override it to be `Some`. Here, we
have it be `Option<Option<u64>>` which allows the `None` override. As
UniFFI doesn't support `Option<Option<..>>`, we work around this via a
dedicated `enum` that is only exposed under the `uniffi` feature.