Make the prober mod public and handle all events (#3)
Forgot to make the prober mod public so that the structs could be used by the Prober.

This also handles all events in ldk node directly so that the the Probe events all get handled correctly while that is happening here. Otherwise, there is a race between event handling in the Prober and event handling here which can cause Probes to hang. This is temporary, next step is to figure out how to serialize those Probe events so that we can do this in the Prober.