bitcoind RPC: Skip validating header chain for newly setup nodes
When we just setup a new node, all chain listeners will still be synced
up to the genesis hash. In order to allow for faster startup, we skip
validating the entirety of the header chain by continuing with the
current best block polled from `bitcoind`.

Note that this doesn't noticeably change our security model as we're
trusting the chain data source, for example to not censor us, anyways.
And, we're of course still validating the PoW for the best block we're
polling.