Pin `lightning-liquidity` in `Cargo.toml`
In the meantime `lightning-liquidity` v0.1.0 was released, which will
have builds of LDK Node v0.4.2 fail as `cargo` by default treats it as
compatible (even though it's not) and tries to
upgrade to the latest version. While users could pin the crate version
back themselves, we here pin it for the time being to fix the v0.4.x
builds.