[LN-1005] Add methods to configure and manipulate the Scorer (#5)
This PR adds scoring parameters to the node configuration. This allows for setting custom scoring parameters which we will use to run experiments. This also adds the methods to import/export/reset the scorer and update its decay parameters. Updating fee parameters requires a restart of the node to update the router