Only skip bumping channel closes for trusted peers
Previously, we'd skip handling any `BumpTransactionEvent` if the
counterparty was trusted. However, this might result in funds getting
stolen by the counterparty in the `ContentiousClaimable` case, as the
counterparty is never expected to spend these back to us. Here, we
therefore begin only skipping the bump events for the `ChannelClose`
variant, and always try to claim funds requiring `HTLCResolution`.