Have `Builder::set_entropy_seed_bytes` take `[u8; 64]` for non-`uniffi`
Historically, we aligned `uniffi` and non-`uniffi` APIs as much as
possible (mostly because there are no rendered docs available for the
latter). As the APIs have diverged a bit by now anyways, we here have
`Builder::set_entropy_seed_bytes` take a `[u8; 64]`, which makes for a
better API in Rust.