Drop `debug_assert` that would have us panic for replayed `PaymentClaimed`s
In cd8958f we changed the internal API behavior of
`PaymentStore::update`. While previously it would return `Ok(true)` the
to-be-updated entry was found in the store, it now returns `Ok(true)` if
not only the entry was found but it was actually updated and
re-persisted. This was an improvement as it allows us to avoid
unnecessary re-persists if nothing changed.

However, there were 1-2 places that implicitly relied on that behavior
for logging purposes which we didn't correctly update to the new
behavior. Unfortunately, one instance in handling `PaymentClaimed`
events actually even `debug_assert`ed on the return value, which lead to
some unnecessary panics in `debug` in case `PaymentClaimed` got
replayed. Here we rectify this by dropping the `debug_assert`.