Retrieve and apply evicted transactions to BDK wallet
With version 2.0, BDK introduced a new `Wallet::apply_evicted_txs` API
that requires that we tell the `Wallet` about any mempool evictions.

Here, we implement this as part of our mempool polling, which should fix
a bug that left the wallet unable to spend funds if
previously-canoncical mempool transactions ended up being evicted.