Stop migrating spendable outputs from pre-v0.3 format
As part of the version v0.3 release we switched to the upstreamed
`OutputSweeper` which slightly changed our serialization format, having
us run a migration step on startup for backwards compatibility ever
since. Here we drop the migration code running on startup, for
simplicity's sake, but also because it's going to be async going forward
and we currently don't have a runtime available on startup (which might
change soon, but still).

As the v0.3 release now well over a year ago, it's very unlikely that
there are any v0.2 (or even v0.3) users left. If there are any affected
users left, they'll first have to upgrade to any version pre-v0.7,
startup, and then upgrade to v0.7 or later.