Drop nexus publishing plugin
Maven Central recently deprecated the Sonatype-style publishing, which
means the nexus publishing gradle plugin we used didn't work anymore. As
Maven Central has yet to release a replacement plugin for gradle, we
simply drop nexus publishing support here and manually upload the
archives in the meantime, which is simple enough.

We also drop the publishing CI jobs that originally aimed to automate
publishing to Maven Central, which we however never came around to use
since we didn't want to fully trust Github CI with publishing binaries
for us.