Refactor move background task fields to `Runtime`
Previously we a) added a new internal `Runtime` API that cleans up our
internal logic and b) added tracking for spawned background tasks to be
able to await/abort them on shutdown.

Here we move the tracking into the `Runtime` object, which will allow us
to easily extend the tracking to *any* spawned tasks in the next step.