Add static invoice support
This commit adds support for using ldk-node as a static invoice server.
When configured as such, the node persists and retrieves invoices from
the configured kv store. Access is guarded by a rate limiter to
prevent overload and mitigate potential DoS attacks.

In this mode, ldk-node also exposes blinded paths that can be shared
with async recipients, allowing them to contact the static invoice
server.

When ldk-node functions as a recipient, it can communicate with the
static invoice server to set up async payments.