Drop `quantity` support from BOLT 12 logic
The BOLT 12 `quantity` field is incredibly unlikely to get any use
in practice (it assumes an online store which places some kind of
static offer for each item on its site, assuming that the merchant
doesn't want structured customer information nor the ability to
sell more than one item in a single order, both of which do not
exist in practice). Worse, supporting it requires an entire UI
flow built around the "quantity" concept, something which is a
nontrivial investment for downstream users of ldk-node.

Because the cost/utility tradeoff isn't nearly worth it, it was
dropped from the main offer-payment API in the upstream `lightning`
crate (requiring a separate
`ChannelManager::pay_for_offer_with_quantity` call).

For the same reason, we simply drop it from our API entirely here.
Absent someone who actually wants to use the `quantity` logic,
there is really no reason to support it.