Re-format all imports
Nightly `rustfmt` allows to auto-group imports on the module level.
While we're not quite convinced to switch to the nightly channel for
this yet (mostly because not all contributors would have the right
nightly version installed on their machines), we here make use of `cargo
+nightly fmt` with some addtional import grouping options as a one-off.
This cleans up our imports for the whole crate and gets us to a
consistent state everywhere.