Remove erroneous `debug_assert` when handling `PaymentForwarded`
Recent LDK changes made `skimmed_fee_msat` a required field of the LSPS
service handler's `payment_forwarded` API, which seemed reasonable given
that the field is available since LDK 0.0.122. However, when updating
LDK Node we introduced a `debug_assert` that checked the field to be
*always* set, which is wrong, as it's only set post-0.0.122 *if* there
was some fee skimmed. Here we fix this oversight.