Fix bitcoind shutdown hang
Previously, the shutdown process could hang when using the bitcoind
backend because the syncing process was not always checking if we had
received the shutdown signal. Now with any future we call during the
sync process we will use a `tokio::select!` to make sure we abort early
if we receive a stop signal.