Use `[patch]` instead of switching all the dependencies
We previously added a bunch of commented-out `rust-lightning`
dependencies in our `Cargo.toml` to be able to easily switch between
`rust-lightning` locations. However, this is exactly what the `[patch]`
command is for, which in particular also allows to patch a dependency
for the whole tree, not only this one project.

Therefore, we move the examples to a commented-out `patch` section.