Prefactor: Move `ChainSource` creation before `Wallet` creation
In the following commits we will use the chain source to poll a best tip
before intializing the listener objects. As a prefactor, we here move
the creation of our onchain wallet after creation of the chain source,
which in turn means we'll need to use the same pattern as for the other
listeners, i.e., not giving the wallet reference to `ChainSource` on
creation but rather handing it in when it's being used at runtime.