Add a `docs.rs` CI job checking documentation builds
We previouly ran into a quiet error that lead to `docs.rs` not rendering
our docs properly, which unfortunately didn't surface until after we
pushed out a releas (thankfully only an RC in this case).

Here, we add a CI job that tests our docs build with exactly the
settings `docs.rs` uses. Additionally, the change also has the benefit
that we now only build docs once rather than for every combiantion in
our workflow matrix, which was a bit overkill.