Make cli output json
Previously we would output the debug format of each of our return times
which worked, but was not very usable. You could not easily pipe results
with cli tools or anything like that. This changes the results to be
output to json to make it easier to use and more similiar to what people
expect.